[#122] Move DebounceSelect as a components
