+++ frontend/src/pages/cases/components/index.js
@@ -179,3 +179,4 @@ export { default as DashboardSensitivityAnalysis } from "./DashboardSensitivityA
 export { default as DashboardScenarioModeling } from "./DashboardScenarioModeling";
 export { default as Questions } from "./Questions";
 export { default as Scenario } from "./Scenario";
+export { default as DebounceSelect } from "./DebounceSelect";
